grabber/requirements.txt
telethon>=1.21.1
python-dotenv>=0.19.0
